refactor(step3): root infra를 common/infra로 이동
