+++ src/main/java/kitchenpos/common/infra/DefaultPurgomalumClient.java
@@ -1,11 +1,10 @@
-package kitchenpos.infra;
+package kitchenpos.common.infra;
 
+import java.net.URI;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
-
-import java.net.URI;
 
 @Component
 public class DefaultPurgomalumClient implements PurgomalumClient {
